feat: replace CodeEditor with EnvEditor for environment variable management
